Refactor parameter type
This change combines the `parameters` and `in` properties of a node into
a single list. This is done by introducing a new type that can either be
a connection to an existing node or a constant value.

It also renames many types to be more clear after taking into account
the new parameter type