Move parameters to dedicated list in graph
When rendering the graph in a GUI, it is easier to have a list of all
the parameters instead of having to step inside each node to find the
parameters. This is because it is useful for the GUI to find all of the
connections it needs to draw between the nodes.